+++ packages/storyboard-listener-console-parallel/src/index.js
@@ -67,7 +67,6 @@ class ParallelConsoleListener {
     const threadId = this.getThread(record);
     if (threadId == null) return;
     if (record.action === 'CLOSED' && record.storyId === threadId) return;
-    // console.log(`fStory=${record.fStory}, parents=${record.parents}, action=${record.action}`)
     const options = timmSet(this.config, 'prevTime', this.prevTime);
     const lines = this.recordToLines(record, options);
     this.prevTime = new Date(record.t);
@@ -120,6 +119,11 @@ class ParallelConsoleListener {
   }
 
   addLineToThread(threadId, line) {
+    // If terminal height is 0 (e.g. Travis), just print the lines out
+    if (!term.height) {
+      console.log(line); // eslint-disable-line
+      return;
+    }
     const thread = this.threads[threadId];
     if (!thread) return;
     thread.lines.push(line);
@@ -147,8 +151,9 @@ class ParallelConsoleListener {
   }
 
   termRefreshAll() {
-    this.termClear();
     const { threads, height: h0 } = this;
+    if (!h0) return;
+    this.termClear();
     const threadIds = Object.keys(threads);
     let len = threadIds.length;
     if (!len) return;
@@ -216,6 +221,7 @@ class ParallelConsoleListener {
   }
 
   resetCursorPos() {
+    if (!term.height) return;
     term.moveTo(1, this.height);
   }
 }